% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi2img.R
\name{midi2img}
\alias{midi2img}
\title{Convert a midi file to an image of its sheet music}
\usage{
midi2img(
  mscore_location = "/Applications/MuseScore 3.app/Contents/MacOS/mscore",
  f
)
}
\arguments{
\item{f}{}
}
\value{

}
\description{
Convert a midi file to an image of its sheet music
}
